Started making some unit tests, and (hand-in-hand) cleaned up the code.
The PopulateOutputConfigurationService had a private method which was
too complex, and therefore warranted its own class
(ParseOutputConfigurationService). I made it so and added a few tests to
get the ball rolling.

(Mental Note) What I Learned:
Private methods require some special treatment when it comes to writing
SOLID code; In this case, I had a class which was performing two logical
operations: reading from a configuration file AND parsing out new
objects from that file. The 'AND' there is key, because it almost always
is an indication that the S in SOLID (Single Responsibility Principal)
is being violated. Since the private parsing method was too complex to
not unit test, I realized the smell: it needed its own service. The
cause of the smell? Being too stingy when making new classes.